chore: add copy-webpack-plugin + rename react/ to components/ :truck:
