+++ components/Icon/styles.styl
