+++ components/Modal/styles.styl
