+++ components/Spinner/styles.styl
