+++ components/Tabs/styles.styl
