+++ package.json
@@ -36,6 +36,7 @@
     "babel-loader": "^7.1.2",
     "babel-preset-env": "^1.1.8",
     "babel-preset-react": "^6.22.0",
+    "copy-webpack-plugin": "^4.0.1",
     "copyfiles": "^1.2.0",
     "css-loader": "^0.28.4",
     "date-fns": "^1.28.5",
